Separate spectra-related codes into a new file
